gcp_mcp_server/pyproject.toml
# pyproject.toml
[project]
name = "gd-mcp-server"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.11.2",
    "google-cloud-bigquery>=3.35.1",
    "google-cloud-storage>=2.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0"
]
